Move DocumentReference Details component to own file
