+++ packages/fhir-client/src/components/PatientDetails/index.tsx
@@ -13,10 +13,8 @@ import { fhirclient } from 'fhirclient/lib/types';
 import { getPatientName, getPath, buildObservationValueString } from '../PatientsList/utils';
 import { resourcesSchema } from '../PatientsList/resourcesSchema';
 import { Dictionary } from '@onaio/utils';
-import { transform } from 'lodash';
 import { IfhirR4 } from '@smile-cdr/fhirts';
-import { format } from 'date-fns';
-import './d.css'; // TODO: should we import this
+import { DocumentReferenceDetails } from '../DocumentReference';
 
 const { Header, Sider, Content } = Layout;
 
@@ -260,6 +258,7 @@ const PatientDetails: React.FC<PatientDetailPropTypes> = (props: PatientDetailPr
                 {Object.keys(resourceTypeMap).map((type: string) => (
                   <Menu.Item
                     key={type}
+                    id={type}
                     onClick={(e) => {
                       setResourceType(e.key as string);
                     }}
@@ -329,65 +328,3 @@ export function ConnectedPatientDetails(props: PatientDetailPropTypes) {
     </QueryClientProvider>
   );
 }
-
-export interface DocumentReferenceDetailsProps {
-  documentResources: IfhirR4.IDocumentReference[];
-}
-
-const DocumentReferenceDetails = (props: DocumentReferenceDetailsProps) => {
-  const { documentResources } = props;
-
-  const getAttachments = (resources: IfhirR4.IDocumentReference[]) => {
-    return resources.map((resource) => {
-      const { id, meta } = resource;
-      const lastUpdated = meta?.lastUpdated
-        ? format(new Date(meta?.lastUpdated), 'yyyy-MM-dd HH:mm:ss')
-        : undefined;
-      const resourceAttachments = resource.content;
-      const parsedAttachments = transform(
-        resourceAttachments,
-        (acc: Dictionary, resAttachment, _) => {
-          const mimeType = resAttachment.attachment.contentType;
-          if (!mimeType) {
-            return false;
-          }
-          if (!acc[mimeType]) {
-            acc[mimeType] = {};
-          }
-          acc[mimeType] = { url: resAttachment.attachment.data };
-        },
-        {}
-      );
-      return { id, lastUpdated, parsedAttachments };
-    });
-  };
-
-  // using a very restrictive check of mime type image/png to identify qr code attachment
-  const qrCodeMimeType = 'image/png';
-  const imgSrcB64Prefix = 'data:image/png;base64,';
-  const hasQrAttachment = (att: ReturnType<typeof getAttachments>[0]) =>
-    !!att.parsedAttachments[qrCodeMimeType];
-
-  return (
-    <div className="documentResource-container">
-      {getAttachments(documentResources).map((attach) => {
-        if (hasQrAttachment(attach)) {
-          const cardProps = {
-            cover: (
-              <img
-                alt={attach.id}
-                src={`${imgSrcB64Prefix} ${attach.parsedAttachments[qrCodeMimeType].url}`}
-              />
-            ),
-          };
-          return (
-            <Card {...cardProps} key={attach.id} className="documentResource-card">
-              <Card.Meta title={attach.id} description={attach.lastUpdated} />
-            </Card>
-          );
-        }
-        return <></>;
-      })}
-    </div>
-  );
-};
